requirements.txt
backcall==0.2.0
cachetools==4.2.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
decorator==5.0.7
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
oauthlib==3.1.0
packaging==20.9
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
pywin32==300
pyzmq==22.0.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tornado==6.1
traitlets==5.0.5
tweepy==3.9.0
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
